format and type hints

+++ flexsim/micro_graph/micro_tensor.py
@@ -10,7 +10,7 @@ class MicroTensor:
     def __init__(self, tensor_shape: Tuple[int, ...]):
         self.tensor_shape = tensor_shape
 
-        self.finish_time = torch.zeros(self.tensor_shape, dtype=torch.int32)
+        self.finish_time = torch.zeros(self.tensor_shape, dtype=torch.float32)
         self.finish_position = torch.zeros(self.tensor_shape, dtype=torch.int16)
 
     def get_finish_time(self):
@@ -29,8 +29,9 @@ class MicroTensor:
         return torch.max(self.finish_time)
 
     def reset(self):
-        self.finish_time = torch.zeros(self.tensor_shape, dtype=torch.int32)
+        self.finish_time = torch.zeros(self.tensor_shape, dtype=torch.float32)
         self.finish_position = torch.zeros(self.tensor_shape, dtype=torch.int16)
+
 
 class MicroTensorNode(Node):
     def __init__(self, graph: Graph, name: str, input_nodes: Tuple[Node, ...], micro_tensor: MicroTensor,
